Support full schema referencing
For when you are storing the whole document in a property and wants to be able to just throw the schema instance in the property and use it right away.

This change will allow you to just do that (see integration tests). **Notice that data added to result will be updated, the stored full document will be available as is only in the document property which will store the full document.**